Subject Line: Add & Implement Menu Item to navigate to PracticeAidActivity page
What Happened:
Added if statement to show EditPracticeAid Menu item only when there is a new RecordActivity to be added/saved.
Comment out if statement because I am not sure if this would confuse the user.
Implement editPracticeAid with an intent to navigate to PracticeAidActivity
Implement PracticeAidActivity to show a message if all Practice Aids have been erased.